restructure profile folder
